Bug 1908693 - Part 1: Propagate MOZ_ environment variables into child processes on iOS, r=glandium!
On desktop platforms, environment variables are implicitly inherited by content
processes, and on Android, custom environment variables are generally set as
part of GeckoThread initialization, which is performed in every process.

As these env vars are not inherited on iOS, we need to manually specify them
for child processes. This patch does this in a fairly hacky way, by enumerating
all environment variables, and copying ones with a `MOZ_` prefix to child
processes as they are created.

We probably want to have a better system for this in the future.

Differential Revision: https://phabricator.services.mozilla.com/D217131