Bug 1908693 - Part 5: Move host tool install logic out of android_device, r=#mozbase-reviewers!,#mach-reviewers!
This logic is used to install the xre host tools, such as a host copy of
xpcshell, which is required when running tests on a remote device. As this will
also be useful for iOS, this logic moves the code to a common place so it can
also be used by iOS in the next part.

Differential Revision: https://phabricator.services.mozilla.com/D225004