Bug 1908693 - Part 6: Add some framework pieces for interacting with iOS devices, r=#mozbase-reviewers!,#mach-reviewers!
This is roughly modeled on the framework used by Android devices, but is
slightly different. This definitely requires a lot of cleanup before it can be
more generally used.

The code is split between mozdevice and mozrunner. While there is some outline
code for running on a real iOS device, most of the core logic is missing, and
would need to be added.

Differential Revision: https://phabricator.services.mozilla.com/D217137