Bug 1908693 - Part 7: Basic support for mochitest-plain on iOS, r=#mozbase-reviewers!,#mach-reviewers!
This patch centers around a hacked up version of runtestsremote.py which
instead uses the helper types from mozdevice.ios (added in the previous part)
to run mochitests on an iOS device. With this change it is possible to run some
basic mochitests on simulator from a local build. There is currently no support
for running on CI.

Tests involving any form of more advanced features such as opening pop-up
windows still are not supported, as the testing browser doesn't support tabs
yet.

Unfortunately, currently a lot of code is duplicated between runtestsremote and
runtestsremoteios. In the future, it might be worthwhile to merge these files
back together, and use a common abstraction instead of having separate
implementations for both platforms.

Differential Revision: https://phabricator.services.mozilla.com/D217138