Bug 1927599 - Part 4: Parent process cannot have child task_t on iOS, r=#ipc-reviewers!
Transferring a task_t over XPC during child process startup appears to cause a
connection failure. It seems likely that transferring a task in this way is a
sandbox violation on-device.

This mostly consists of changing ifdefs associated with the task_t object to be
macOS-specific instead. iOS simulator builds will stop working after this
change until the IPC backend is changed.

Differential Revision: https://phabricator.services.mozilla.com/D227102