Bug 1927599 - Part 7: Use XPC as an IPC backend on iOS, r=#ipc-reviewers!
This approach, especially in its current form, has quite a bit of overhead, but
should help us get some prototyping done for iOS. In the future, we'll probably
want to directly use mach ports, either with a kevent-based message loop (if
that syscall is unblocked) or with a libdispatch queue replacing our IPC IO
Thread.

This also required shimming out the process watcher, which also used the
libevent pump.

Differential Revision: https://phabricator.services.mozilla.com/D227105