Bug 1931951 - Share default prefs between android and iOS, r=#geckoview-reviewers!
This allows more logic to be handled in a common place between these two implementations.

I expect that more of the prefs previously set in the android code are actually
android specific, and shouldn't be set on iOS, but for now this shares almost
everything, as the backends are theoretically very similar.