Bug 1953595 - Part 1: Refactor EventDispatcher to be more embedder-agnostic, r=#geckoview-reviewers!
In this patch, the logic for translating data types between JS/Java has been
moved to a common abstraction called `GeckoViewData`. This interface consists
of a `GeckoViewDataSink` and a `GeckoViewDataSource`, along with a method to
drive one from the other.

This allows the JS logic to be separated from the Java logic, which will make
an Objective-C/Swift backend for iOS much easier to implement and continue to
support. In the future, we may also want to add a C++ backend to make code
which needs to dispatch events to the bus from C++ easier to use.

In addition, this patch also refactors the C++/JS side of EventDispatcher which
into a new EventDispatcherBase class, which has no dependencies on jni or
Android thanks to passing around `GeckoViewDataSource` references. This will
allow the iOS implementation in part 2 to be much smaller and share more code
with the Android implementation.

When refacotring the EventDispatcher logic, I also noticed some small
simplifications (such as using `nsTObserverArray` rather than the hand-rolled
observer logic from the previous version of the patch), which have been
applied. The code was also updated to use proper thread-safety annotations.

Differential Revision: https://phabricator.services.mozilla.com/D241287