Bug 1953595 - Part 2: Add EventDispatcher and GeckoViewData implementation for iOS, r=glandium!
With the new EventDispatcher implementation and abstractions added for Android
in earlier parts, adding an EventDispatcher implementation for iOS is fairly
straightforward.

This patch adds a GeckoViewData{Source,Sink} implementation for the
CoreFoundation data structures, which support "toll-free bridging" (i.e. direct
casting) to/from the Objective-C Foundation types (and therefore Swift data
structures).

In addition, the stub EventDispatcher class is changed to inherit from the new
common EventDispatcherBase class, meaning the bulk of the logic is fully shared
with the Android implementation.

The Swift part of this change will be added in bug 1945252, alongside the XCode
project, as there is unfortunately nowhere to build Swift code in the tree
until that project is added.

Differential Revision: https://phabricator.services.mozilla.com/D242854