Bug 1953595 - Part 2: Add EventDispatcher implementation for iOS, r=glandium!
In this patch, the stub EventDispatcher class is changed to inherit from the
new common EventDispatcherBase class, which aims to share some logic with the
Android implementation.

The type conversion logic is written using CoreFoundation data structures,
which support "toll-free bridging" (i.e. direct casting) to/from the
Objective-C Foundation types (and therefore Swift data structures).

This is done despite `EventDispatcher` currently being implemented in
Objective-C++ as it provides a slightly more C++-friendly interface, and is
easier to manage the lifetimes of objects using CFTypeRefPtr with. This also
allows us to move the type conversion logic to a separate file in the future if
it is needed elsewhere.

The Swift part of this change will be added in bug 1945252, alongside the XCode
project, as there is unfortunately nowhere to build Swift code in the tree
until that project is added.

Differential Revision: https://phabricator.services.mozilla.com/D242854