Bug 901050 - Part 2: Add mach port watching support to the IO thread, r=#ipc-reviewers!
This adds a new MessagePump implementation, MessagePumpKqueue. This new
implementation is interface-compatible with the existing MessagePumpLibevent,
but adds support for also watching mach ports.

This version of MessagePumpKqueue was originally ported from chromium in
~2022. The code was backported to work with the older version of MessageLoop
used in our fork of Chromium, and use XPCOM data structures where needed.

This implementation of MessagePumpKqueue uses mach ports heavily, including for
wake-up events, and is not compatible with non-Darwin BSDs which will continue
to use libevent.

Differential Revision: https://phabricator.services.mozilla.com/D251393