Bug 901050 - Part 7: Use the SEND_POSSIBLE notification for mach_msg timeouts, r=#ipc-reviewers!
This patch adds support to message_pump_kqueue for watching for SEND_POSSIBLE
and DEAD_NAME notifications for send rights from the kernel.

This is done in the message pump, rather than in the IPC::Channel directly, as
failing to process a queued DEAD_NAME notification from the kernel (e.g. due to
destroying a receive right with queued messages) can lead to a port leak. By
using a shared listener port in the message pump, this leak can be avoided.

Differential Revision: https://phabricator.services.mozilla.com/D254420